requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
scipy>=1.11.0

# Evaluation and Metrics
evaluate>=0.4.1
scikit-learn>=1.3.0
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Monitoring and Logging
wandb>=0.16.0
tensorboard>=2.15.0
tqdm>=4.66.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
sentencepiece>=0.1.99
protobuf>=4.25.0

